pyproject.toml
[project]
name = "vectordb-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.9.0",
    "sentence-transformers>=2.6.1",
]
